% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmedsl-package.R
\docType{package}
\name{ggmedsl-package}
\alias{ggmedsl}
\alias{ggmedsl-package}
\title{ggmedsl: Create ggplot2 Graphics Using MEDSL Colors and Branding}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This package helps you create ggplot2 graphics Using MEDSL colors and branding.
}
\author{
\strong{Maintainer}: Cameron Wimpy \email{wimpy@mit.edu}

}
\keyword{internal}
