% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsl_map.R
\name{medsl_map}
\alias{medsl_map}
\title{Add MEDSL theme to ggplot-based map objects.}
\usage{
medsl_map()
}
\description{
This function allows you to add the MEDSL theme to your ggplot-based maps.
}
\keyword{medsl_map}
