% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsl_plot.R
\name{medsl_plot}
\alias{medsl_plot}
\title{Add MEDSL theme to ggplot graphs.}
\usage{
medsl_plot()
}
\description{
This function allows you to add the MEDSL theme to your ggplot graphs.
}
\examples{
\dontrun{
ggplot(mtcars, aes(mpg, disp, color = factor(carb))) +
geom_point(size = 3) +
medsl_plot() +
}

}
\keyword{medsl_plot}
