% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_medsl.R
\name{scale_fill_medsl}
\alias{scale_fill_medsl}
\title{Add MEDSL color fill to ggplot graphs.}
\usage{
scale_fill_medsl()
}
\description{
This function allows you to add the MEDSL color fill to your ggplot graphs.
}
\examples{

ggplot(mtcars, aes(mpg, disp, color = factor(carb))) +
geom_point(size = 3) +
ggmedsl() +
scale_fill_medsl()
}
\keyword{scale_fill_medsl}
