% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmedsl-package.R
\docType{package}
\name{ggmedsl-package}
\alias{ggmedsl}
\alias{ggmedsl-package}
\title{ggmedsl: Create ggplot2 Graphics Using MEDSL Colors and Branding.}
\description{
This package simplifies the creation of ggplot2 graphics using MEDSL branding and colors.
}
\author{
\strong{Maintainer}: Cameron Wimpy \email{wimpy@mit.edu}

}
\keyword{internal}
