% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_medsl.R
\name{scale_color_medsl}
\alias{scale_color_medsl}
\title{Add MEDSL colors to ggplot graphs.}
\usage{
scale_color_medsl()
}
\description{
This function allows you to add the MEDSL colors to your ggplot graphs.
}
\examples{
ggplot(mtcars, aes(mpg, disp, color = factor(carb))) +
geom_point(size = 3) +
ggmedsl() +
scale_color_medsl()
}
\keyword{scale_color_medsl}
